Convert all old database to new one

+++ BookBook/Models/ProductByType.cs
@@ -9,7 +9,7 @@ namespace BookBook.Models
     {
         public int id { get; set; }
         public string name { get; set; }
-        public double price { get; set; }
+        public int price { get; set; }
         public string image { get; set; }
     }
 }
